% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{assess_predictions}
\alias{assess_predictions}
\title{Prepare data.frames for multiple model assessment visualizations}
\usage{
assess_predictions(yhat, y, yhat_label = "Predictions")
}
\arguments{
\item{yhat}{A vector of predicted outcome values}

\item{y}{A vector of observed outcome values}

\item{yhat_label}{Optional label to describe the model output. Useful when
multiple models will be compared in a single visualization. Defaults to
"Predictions".}
}
\value{
A named list containing 4 \code{data.frame}s: "res" (containing
predicted and observed outcome values), "roc" (ROC curve), "pr"
(precision-recall curve), and "acc" (accuracy curve).
}
\description{
Prepare data.frames for multiple model assessment visualizations
}
