% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_scale_norm.R
\name{distByRows}
\alias{distByRows}
\title{Estimate the record-level distance from \code{validX} to \code{trainX}}
\usage{
distByRows(
  normF = c("L1", "L2", "L3", "L4", "L5", "L6", "max"),
  validX,
  trainX,
  keepX,
  verbose = FALSE
)
}
\arguments{
\item{normF}{a character string representing one of the following functions:
\itemize{
\item \code{L1}: Manhattan aka 1-norm or L1-norm (default)
\item \code{L2}: Euclidean aka 2-norm or L2-norm
\item \code{L3}: Minkowski aka 3-norm or L3-norm
\item \code{L4}: Minkowski aka 4-norm or L4-norm
\item \code{L5}: Minkowski aka 5-norm or L5-norm
\item \code{L6}: Minkowski aka 6-norm or L6-norm
\item \code{max}: Chebyshev aka infinity norm
}}

\item{validX}{A \code{matrix}, \code{data.frame}, or numeric vector
containing one or more observations to which \code{trainX} should be
normalized (typically a validation/test set).}

\item{trainX}{A \code{data.frame} (typically containing the training set) to
be normalized.}

\item{keepX}{Character vector containing names of retained features.}

\item{verbose}{Currently unused. A boolean indicating whether diagnostic
messages should be printed. Defaults to \code{FALSE}.}
}
\value{
A \code{matrix}
}
\description{
Estimate the record-level distance from \code{validX} to \code{trainX}
}
