% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{distKernelWeights}
\alias{distKernelWeights}
\title{Estimate distances in "neighborhoods" using specified features}
\usage{
distKernelWeights(
  keepX,
  trainX,
  testX,
  neighbor_library,
  save_dist = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{keepX}{Character vector containing names of retained features.}

\item{trainX}{A \code{data.frame} (typically containing the training/tuning
set) to be transformed.}

\item{testX}{A \code{data.frame} containing one or more observations to
transform \code{trainX} against (typically a validation/test set).}

\item{neighbor_library}{A \code{data.frame} containing columns "norm",
"kernel", and "window", with one row for each unique combination of settings
that should be applied. Column "norm" should contain valid norm function
names (see the \code{normF} argument of \code{\link{distByRows}} for
options). Column "kernel" should contain valid kernel function names (see the
\code{kernel} argument of \code{\link[kedd]{kernel.fun}} for options). Column
"window" should contain valid window sizes (see the \code{windowProp}
argument of \code{\link{scaleRowDists}} for details).}

\item{save_dist}{A boolean indicating whether to retain the raw distance
matrices in the resulting \code{data.frame} in an additional (list) column
named "dist". Defaults to \code{FALSE}.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
The \code{neighbor_library} \code{data.frame} supplemented with at
least two columns: "windowFmtd" (a formatted string version of the supplied
column "window") and "kernel_weights". Each row's "kernel_weights" contains a
sparse matrix of distances with number of rows equal to the number of rows in
\code{testX} and the number of and number of columns equal to the number of
rows in \code{trainX}.
}
\description{
Estimate distances in "neighborhoods" using specified features
}
\examples{
neighbor_lib <- tidyr::expand_grid(norm = c("L1", "L2"),
                                   kernel = c("uniform", "tricube", "epanechnikov"),
                                   window = c(0.1, 0.2, 0.3))
res <- distKernelWeights(keepX = c("qsec", "cyl", "disp", "mpg", "carb", "drat", "hp", "wt"),
                         trainX = mtcars[1:28, -which(colnames(mtcars) \%in\% "vs")],
                         testX = mtcars[29:32, -which(colnames(mtcars) \%in\% "vs")],
                         neighbor_library = neighbor_lib)
}
