% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitPredictLibSL}
\alias{fitPredictLibSL}
\title{Alternative Super Learner Prediction Function}
\usage{
fitPredictLibSL(
  Y,
  X,
  newX = NULL,
  family = gaussian(),
  SL.library,
  id = NULL,
  verbose = FALSE,
  saveFitLibrary = FALSE,
  obsWeights = NULL
)
}
\arguments{
\item{Y}{Numeric vector containing observed outcome values for the training
set.}

\item{X}{A \code{data.frame} containing predictor values for the training
set.}

\item{newX}{Optional \code{data.frame} containing predictor values for the
test set.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial". Defaults to
\code{\link{gaussian}}.}

\item{SL.library}{A character vector of prediction algorithm names or a list
containing these character vectors.}

\item{id}{Optional numeric vector containing cluster IDs.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}

\item{saveFitLibrary}{Boolean indicating whether to retain the fit object for
each algorithm in the library in the returned result. Defaults to
\code{FALSE}.}

\item{obsWeights}{Optional numeric vector containing observation weights.}
}
\value{
A list. See \code{\link[SuperLearner]{SampleSplitSuperLearner}}
for details.
}
\description{
This function is a refactored and stripped down version of
\code{\link[SuperLearner]{SampleSplitSuperLearner}}.
}
