% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{GridLearner}
\alias{GridLearner}
\title{De-Supered SuperLearner}
\usage{
GridLearner(
  Y,
  X,
  newX = NULL,
  family = gaussian(),
  SL.library,
  id = NULL,
  verbose = FALSE,
  control = list(),
  obsWeights = NULL,
  env = parent.frame()
)
}
\arguments{
\item{Y}{A numeric vector of outcome values for only the training set.}

\item{X}{A \code{data.frame} of selected predictor values for only the
training set.}

\item{newX}{A single-row \code{data.frame} of selected predictor values for
only the current validation observation.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{SL.library}{A list containing character vectors where the first element
is a screening algorithm and the second is a prediction algorithm. See
\code{\link[SuperLearner]{SuperLearner}}'s \code{SL.library} argument for
more details.}

\item{id}{Optional cluster identification variable.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}

\item{control}{A list containing two named elements: "saveFitLibrary" (a
boolean indicating whether to retain the fit object for each algorithm in the
library in the returned result) and "trimLogit" (for the NNloglik combination
method, the level to which to truncate the logit transformation). See
\code{\link[SuperLearner]{SuperLearner.control}} for details.}

\item{obsWeights}{A sparse vector of distances for only the current
validation observation with length equal to the number of rows in \code{X}.}

\item{env}{Environment containing the prediction functions.}
}
\value{
A \code{list} of class \code{SuperLearner}.
}
\description{
The original \code{\link[SuperLearner]{SuperLearner}} was created by Eric
Polley in 2011. This function is a modified for purpose version of
\code{\link[SuperLearner]{SuperLearner}} which removes the cross-validation
and simply returns the predictions from each library algorithm.
}
