% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_scale_norm.R
\name{normalizeFeatures}
\alias{normalizeFeatures}
\title{Normalize a dataset to a single observation}
\usage{
normalizeFeatures(obs, X, keepX = colnames(X))
}
\arguments{
\item{obs}{A \code{data.frame} containing a single record/row (typically a
validation/test set observation) to which \code{X} will be normalized.}

\item{X}{A \code{data.frame} (typically containing the training set) to be
normalized.}

\item{keepX}{Character vector containing names of retained features.}
}
\value{
A \code{data.frame}
}
\description{
Normalize a dataset \code{X} to a single observation \code{obs} by
subtracting the observation's values and dividing by the L1 norm.
}
