% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{predictAllYALL}
\alias{predictAllYALL}
\title{Use a single set of tuning parameters to predict outcomes in the holdout set}
\usage{
predictAllYALL(
  newX,
  y_all,
  x_all,
  family,
  predSummaryRow,
  libraryCVSLFeatSel,
  nCVSLFeatSelFolds,
  verbose = FALSE
)
}
\arguments{
\item{newX}{A \code{data.frame} of predictor values for the entire holdout
set.}

\item{y_all}{A numeric vector of outcome values for the entire tuning set.}

\item{x_all}{A \code{data.frame} of predictor values for the entire tuning
set.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{predSummaryRow}{A single-row \code{data.frame} containing at least
columns "combo_method", "selector", "k", "norm", "kernel", "window",
"predictor", "screener"}

\item{libraryCVSLFeatSel}{A list of character vectors of length 2, each
containing a screener algorithm and a prediction algorithm, to be used in
global feature selection by \code{\link[SuperSelector]{cvSLFeatureSelector}}.
See \code{\link[SuperLearner]{SuperLearner}}.}

\item{nCVSLFeatSelFolds}{Numeric vector of length 1 or 2. Used to specify
the number of cross-validation folds for global feature selection in
\code{\link[SuperSelector]{cvSLFeatureSelector}}.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A named list with elements "predict" (a \code{list}) and "screen" (a
\code{data.frame}.
}
\description{
Use a single set of tuning parameters to predict outcomes in the holdout set
}
