% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{predictOneYALL}
\alias{predictOneYALL}
\title{Use a single set of tuning parameters to predict outcomes for a single record in the holdout set}
\usage{
predictOneYALL(
  newX,
  y_all,
  x_all,
  family,
  neighbor_control,
  predict_control,
  keepX,
  screenRow,
  verbose = FALSE
)
}
\arguments{
\item{newX}{A single-row \code{data.frame} of predictor values for one record
in the holdout set.}

\item{y_all}{A numeric vector of outcome values for the entire tuning set.}

\item{x_all}{A \code{data.frame} of predictor values for the entire tuning
set.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{neighbor_control}{A named list with one element "library" containing a
single-row \code{data.frame} with columns "norm", "kernel", and "window"
specifying tuning parameters to be used when calling
\code{\link{distKernelWeights}}.}

\item{predict_control}{A named list with one element "SL.library" containing
a single-element list, which in turn contains a character vector listing the
prediction and screening functions to be used when calling
\code{fitPredictLibSL}.}

\item{keepX}{A character vector of features in \code{x_all} and \code{newX}
to retain.}

\item{screenRow}{The "whichVariable" element of the result from calling
\code{\link[SuperSelector]{cvSLFeatureSelector}}.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
Use a single set of tuning parameters to predict outcomes for a single record in the holdout set
}
