% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{pretty_cvauc_ci}
\alias{pretty_cvauc_ci}
\title{Cross-validated AUC estimate with confidence interval as a formatted string}
\usage{
pretty_cvauc_ci(yhat, y, folds = NULL, confidence = 0.95)
}
\arguments{
\item{yhat}{A vector of predicted outcome values}

\item{y}{A vector of observed outcome values}

\item{folds}{Vector of fold IDs. If \code{NULL}, the standard error will be
estimated as though there were a single fold (no cross-validation).}

\item{confidence}{Confidence level; defaults to \code{0.95}}
}
\value{
A character string
}
\description{
Cross-validated AUC estimate with confidence interval as a formatted string
}
