% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_scale_norm.R
\name{rowNorms}
\alias{rowNorms}
\title{Calculate the magnitude of row-vectors}
\usage{
rowNorms(x, f = c("L1", "L2", "L3", "L4", "L5", "L6", "max"))
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}}

\item{f}{a character string representing one of the following functions:
\itemize{
\item \code{L1}: Manhattan aka 1-norm or L1-norm (default)
\item \code{L2}: Euclidean aka 2-norm or L2-norm
\item \code{L3}: Minkowski aka 3-norm or L3-norm
\item \code{L4}: Minkowski aka 4-norm or L4-norm
\item \code{L5}: Minkowski aka 5-norm or L5-norm
\item \code{L6}: Minkowski aka 6-norm or L6-norm
\item \code{max}: Chebyshev aka infinity norm
}}
}
\value{
a numeric vector
}
\description{
Calculate the magnitude of row-vectors
}
