% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{scaleDistForKernel}
\alias{scaleDistForKernel}
\title{Scale a numeric vector according to a window size}
\usage{
scaleDistForKernel(d, windowProp = 0.2, verbose = FALSE)
}
\arguments{
\item{d}{A numeric vector of distances. In a typical use case, \code{d} will
have length equal to the number of training set observations.}

\item{windowProp}{Window size, as a proportion (0, 1]. Defaults to
\code{0.20}.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A numeric vector with length equal to the length of \code{d}.
}
\description{
Scale a numeric vector according to a window size
}
