% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{scaleRowDists}
\alias{scaleRowDists}
\title{Row-wise scaling according to a window size}
\usage{
scaleRowDists(windowProp, d, verbose = FALSE)
}
\arguments{
\item{windowProp}{Window size, as a proportion (0, 1].}

\item{d}{a matrix with rows corresponding to test set observations and
columns corresponding to training set observations}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
a matrix with rows corresponding to test set observations and columns
corresponding to training set observations
}
\description{
Row-wise scaling according to a window size
}
