% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_screeners.R
\name{screen.corRank3.wgtd}
\alias{screen.corRank3.wgtd}
\title{Shortcut function to \code{screen.wgtd.corRank} with \code{k = 3}}
\usage{
screen.corRank3.wgtd(..., k = 3)
}
\arguments{
\item{...}{Passed through to \code{\link[SLScreenExtra]{screen.wgtd.corRank}}}

\item{k}{Minimum number of features to select. Set to \code{3}.}
}
\value{
A logical vector
}
\description{
Shortcut function to \code{screen.wgtd.corRank} with \code{k = 3}
}
