% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_screeners.R
\name{screen.corRank4.wgtd}
\alias{screen.corRank4.wgtd}
\title{Shortcut function to \code{screen.wgtd.corRank} with \code{k = 4}}
\usage{
screen.corRank4.wgtd(..., k = 4)
}
\arguments{
\item{...}{Passed through to \code{\link[SLScreenExtra]{screen.wgtd.corRank}}}

\item{k}{Minimum number of features to select. Set to \code{4}.}
}
\value{
A logical vector
}
\description{
Shortcut function to \code{screen.wgtd.corRank} with \code{k = 4}
}
