% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_screeners.R
\name{screen.corRank5.wgtd}
\alias{screen.corRank5.wgtd}
\title{Shortcut function to \code{screen.wgtd.corRank} with \code{k = 5}}
\usage{
screen.corRank5.wgtd(..., k = 5)
}
\arguments{
\item{...}{Passed through to \code{\link[SLScreenExtra]{screen.wgtd.corRank}}}

\item{k}{Minimum number of features to select. Set to \code{5}.}
}
\value{
A logical vector
}
\description{
Shortcut function to \code{screen.wgtd.corRank} with \code{k = 5}
}
