% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_screeners.R
\name{screen.corRank6.wgtd}
\alias{screen.corRank6.wgtd}
\title{Shortcut function to \code{screen.wgtd.corRank} with \code{k = 6}}
\usage{
screen.corRank6.wgtd(..., k = 6)
}
\arguments{
\item{...}{Passed through to \code{\link[SLScreenExtra]{screen.wgtd.corRank}}}

\item{k}{Minimum number of features to select. Set to \code{6}.}
}
\value{
A logical vector
}
\description{
Shortcut function to \code{screen.wgtd.corRank} with \code{k = 6}
}
