% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_wrappers.R
\name{SL.glm.mean}
\alias{SL.glm.mean}
\title{A wrapper for the mean via \code{glm}.}
\usage{
SL.glm.mean(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{Numeric vector containing observed outcome values for the training
set.}

\item{X}{A \code{data.frame} containing predictor values for the training
set.}

\item{newX}{A \code{data.frame} containing predictor values for the test set.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{obsWeights}{Optional numeric vector of weights with length equal to
the length of \code{Y} and the number of rows in \code{X}.}

\item{...}{Currently unused.}
}
\value{
A numeric vector of predictions.
}
\description{
A wrapper for the mean via \code{glm}.
}
