% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{summarizeYALL}
\alias{summarizeYALL}
\title{Summarize "predict" element of YALL model results}
\usage{
summarizeYALL(df)
}
\arguments{
\item{df}{The \code{data.frame} stored in the "predict" element of results
from a YALL model}
}
\value{
A \code{data.frame} summary
}
\description{
Summarize "predict" element of YALL model results
}
