% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{SuperLearnerLOO}
\alias{SuperLearnerLOO}
\title{Wrapper function for \code{GridLearner}}
\usage{
SuperLearnerLOO(
  X,
  Y,
  W,
  newX,
  newY,
  family,
  SL.library,
  control,
  verbose = FALSE
)
}
\arguments{
\item{X}{A \code{data.frame} of selected predictor values for only the
training set.}

\item{Y}{A numeric vector of outcome values for only the training set.}

\item{W}{A sparse vector of distances for only the current validation
observation with length equal to the number of rows in \code{X}.}

\item{newX}{A single-row \code{data.frame} of selected predictor values for
only the current validation observation.}

\item{newY}{A numeric outcome value for only the current validation
observation.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{SL.library}{A list containing character vectors where the first element
is a screening algorithm and the second is a prediction algorithm. See
\code{\link[SuperLearner]{SuperLearner}}'s \code{SL.library} argument for
more details.}

\item{control}{A list containing two named elements: "saveFitLibrary" (a
boolean indicating whether to retain the fit object for each algorithm in the
library in the returned result) and "trimLogit" (for the NNloglik combination
method, the level to which to truncate the logit transformation). See
\code{\link[SuperLearner]{SuperLearner.control}} for details.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A \code{list} of class \code{SuperLearner}.
}
\description{
Wrapper function for \code{GridLearner}
}
