% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{SuperLearnerLOOMultiValid}
\alias{SuperLearnerLOOMultiValid}
\title{For one validation set row and each set of parameters, estimate library predictions}
\usage{
SuperLearnerLOOMultiValid(
  validIdx,
  validRows,
  X,
  Y,
  weightDF,
  family,
  SL.library,
  control,
  verbose = FALSE
)
}
\arguments{
\item{validIdx}{Index within \code{validRows} representing the 'current'
validation observation.}

\item{validRows}{A numeric vector of row numbers within \code{X} (and indices
within \code{Y}) to use as a validation set within the current fold.}

\item{X}{A \code{data.frame} of predictor values for the entire tuning set.}

\item{Y}{A numeric vector of outcome values for the entire tuning set.}

\item{weightDF}{A \code{data.frame} containing at least columns "keep_names"
(where each value contains a character vector of retained features) and
"kernel_weights" (where each value contains a sparse matrix of distances with
number of rows equal to length of \code{validRows} and the number of and
number of columns equal to the number of rows in \code{X} minus the length of
\code{validRows}).}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{SL.library}{A list containing character vectors where the first element
is a screening algorithm and the second is a prediction algorithm. See
\code{\link[SuperLearner]{SuperLearner}}'s \code{SL.library} argument for
more details.}

\item{control}{A list containing two named elements: "saveFitLibrary" (a
boolean indicating whether to retain the fit object for each algorithm in the
library in the returned result) and "trimLogit" (for the NNloglik combination
method, the level to which to truncate the logit transformation). See
\code{\link[SuperLearner]{SuperLearner.control}} for details.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
For one validation set row and each set of parameters, estimate library predictions
}
