% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{SuperLearnerLOOValid}
\alias{SuperLearnerLOOValid}
\title{For one validation set row and one set of parameters, estimate library predictions}
\usage{
SuperLearnerLOOValid(
  X,
  W,
  Y,
  validIdx,
  validRows,
  family,
  SL.library,
  control,
  verbose = FALSE
)
}
\arguments{
\item{X}{A \code{data.frame} of selected predictor values for the entire
tuning set.}

\item{W}{A sparse matrix of distances with number of rows equal to length of
\code{validRows} and the number of and number of columns equal to the number
of rows in \code{X} minus the length of \code{validRows} (number of training
set observations).}

\item{Y}{A numeric vector of outcome values for the entire tuning set.}

\item{validIdx}{Index within \code{validRows} representing the 'current'
validation observation.}

\item{validRows}{A numeric vector of row numbers within \code{X} (and indices
within \code{Y}) to use as a validation set within the current fold.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{SL.library}{A list containing character vectors where the first element
is a screening algorithm and the second is a prediction algorithm. See
\code{\link[SuperLearner]{SuperLearner}}'s \code{SL.library} argument for
more details.}

\item{control}{A list containing two named elements: "saveFitLibrary" (a
boolean indicating whether to retain the fit object for each algorithm in the
library in the returned result) and "trimLogit" (for the NNloglik combination
method, the level to which to truncate the logit transformation). See
\code{\link[SuperLearner]{SuperLearner.control}} for details.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
For one validation set row and one set of parameters, estimate library predictions
}
