% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{the_prediction_function}
\alias{the_prediction_function}
\title{Prediction function used by \code{GridLearner}}
\usage{
the_prediction_function(
  index,
  lib,
  Y,
  dataX,
  newX,
  whichScreen,
  family,
  id,
  obsWeights,
  verbose,
  control,
  libraryNames,
  env,
  fitLibEnv
)
}
\arguments{
\item{index}{Index of prediction algorithm to use in \code{lib}}

\item{lib}{Prediction algorithm (output of
\code{SuperLearner:::.createLibrary}).}

\item{Y}{A numeric vector of outcome values for only the training set.}

\item{dataX}{A \code{data.frame} of selected predictor values for only the
training set.}

\item{newX}{A single-row \code{data.frame} of selected predictor values for
only the current validation observation.}

\item{whichScreen}{A boolean vector indicating which columns in \code{dataX}
and \code{newX} to use in the prediction.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{id}{Optional cluster identification variable.}

\item{obsWeights}{A sparse vector of distances for only the current
validation observation with length equal to the number of rows in
\code{dataX}.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}

\item{control}{A list containing two named elements: "saveFitLibrary" (a
boolean indicating whether to retain the fit object for each algorithm in the
library in the returned result) and "trimLogit" (for the NNloglik combination
method, the level to which to truncate the logit transformation). See
\code{\link[SuperLearner]{SuperLearner.control}} for details.}

\item{libraryNames}{A character string label describing the screening and
prediction function combination.}

\item{env}{Environment containing the prediction functions.}

\item{fitLibEnv}{Environment containing fitLibrary.}
}
\value{
A numeric vector of predicted outcome values.
}
\description{
Prediction function used by \code{GridLearner}
}
