% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{trainYALLfold}
\alias{trainYALLfold}
\title{Train the local learning algorithm on a single CV fold}
\usage{
trainYALLfold(
  y_all,
  x_all,
  fold_num,
  trainRows,
  validRows,
  family,
  cvSLfeatsel_control,
  neighbor_control,
  predict_control,
  save_dist = FALSE,
  save_fit_library = TRUE,
  model_keep = c("everything"),
  verbose = FALSE
)
}
\arguments{
\item{y_all}{A numeric vector of outcome values for the entire
tuning set.}

\item{x_all}{A \code{data.frame} of predictor values for the entire
tuning set.}

\item{fold_num}{A numeric value specifying the CV fold ID.}

\item{trainRows}{A numeric vector of row numbers within \code{x_all} (and
indices within \code{y_all}) to use for training for \code{fold_num}.}

\item{validRows}{A numeric vector of row numbers within \code{x_all} (and
indices within \code{y_all}) to use for validation for \code{fold_num}.}

\item{family}{A character string or function describing the error
distribution. Should be one of "gaussian" or "binomial".}

\item{cvSLfeatsel_control}{A named list containing tuning options for
\code{\link[SuperSelector]{cvSLFeatureSelector}}: "method", "SL.library",
"selector.library", "nFolds".}

\item{neighbor_control}{A named list containing a single tuning option
element for \code{link{distKernelWeights}}: "library". See the
\code{neighbor_library} argument of \code{link{distKernelWeights}} for
details.}

\item{predict_control}{A named list containing tuning options passed to
\code{\link{GridLearner}}: "SL.library" and "trimLogit".}

\item{save_dist}{A boolean indicating whether to retain the distance values
in the result. Passed to \code{\link{distKernelWeights}}. Defaults to
\code{FALSE}.}

\item{save_fit_library}{A boolean passed to \code{\link{GridLearner}} to
indicate whether to retain the fit object for each algorithm in the library
in the returned result. Defaults to \code{TRUE}. Note that a \code{TRUE}
setting can result in a very large returned result.}

\item{model_keep}{If \code{save_fit_library} is true, a character vector
indicating which, if any, "disposable" elements of the model fit objects
to retain. Defaults to \code{"everything"}. See \code{\link[strip]{strip}}
for details.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A named list with elements "predict", "screen", "screen_full", and
"screen_which".
}
\description{
\code{trainYALLfold} is the main interface to training the local learning
algorithm within a single cross-validation fold.
}
