% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{weightNeighbors}
\alias{weightNeighbors}
\title{Apply a kernel function to a numeric vector}
\usage{
weightNeighbors(scaledDist, kernelFunction, verbose = FALSE)
}
\arguments{
\item{scaledDist}{A numeric vector}

\item{kernelFunction}{A character string denoting the kernel function to be
used. Defaults to "gaussian". See \code{\link[kedd]{kernel.fun}} for details.}

\item{verbose}{Currently unused. A boolean indicating whether diagnostic
messages should be printed. Defaults to \code{FALSE}.}
}
\value{
A numeric vector with length equal to the length of
\code{scaledDist}.
}
\description{
Apply a kernel function to a numeric vector
}
