% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{weightRowNeighbors}
\alias{weightRowNeighbors}
\title{Row-wise application of a kernel function}
\usage{
weightRowNeighbors(kernelFunction, d, verbose = FALSE)
}
\arguments{
\item{kernelFunction}{A character string denoting the kernel function to be
used. See \code{\link[kedd]{kernel.fun}} for details.}

\item{d}{a matrix with rows corresponding to test set observations and
columns corresponding to training set observations}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
a matrix with rows corresponding to test set observations and
columns corresponding to training set observations
}
\description{
Row-wise application of a kernel function
}
