% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weights.R
\name{distKernelWeights}
\alias{distKernelWeights}
\title{Short description}
\usage{
distKernelWeights(
  keepX,
  trainX,
  testX,
  neighbor_library,
  save_dist = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{keepX}{Character vector containing names of retained features.}

\item{trainX}{A \code{data.frame} (typically containing the training/tuning
set) to be transformed.}

\item{testX}{A \code{data.frame} containing one or more observations to
transform \code{trainX} against (typically a validation/test set).}

\item{neighbor_library}{A \code{data.frame} containing columns "norm",
"kernel", and "window", with one row for each unique combination of settings
that should be applied. Column "norm" should contain valid norm function
names (see the \code{normF} argument of \code{\link{distByRows}} for
options). Column "kernel" should contain valid kernel function names (see the
\code{kernel} argument of \code{\link[kedd]{kernel.fun}} for options). Column
"window" should contain valid window sizes (see the \code{windowProp}
argument of \code{\link{scaleRowDists}} for details).}

\item{save_dist}{A boolean indicating whether to retain the distance values
in the result. Defaults to \code{FALSE}.}

\item{verbose}{A boolean indicating whether diagnostic messages should be
printed. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
Short description
}
